+++ src/euroeval/dataset_configs/danish.py
@@ -2,7 +2,18 @@
 
 from ..data_models import DatasetConfig
 from ..languages import DANISH
-from ..tasks import COMMON_SENSE, EUROPEAN_VALUES, KNOW, LA, MCRC, NER, RC, SENT, SUMM
+from ..tasks import (
+    COMMON_SENSE,
+    EUROPEAN_VALUES,
+    KNOW,
+    LA,
+    LOGIC,
+    MCRC,
+    NER,
+    RC,
+    SENT,
+    SUMM,
+)
 
 ### Official datasets ###
 
@@ -144,5 +155,25 @@ WINOGRANDE_DA_CONFIG = DatasetConfig(
     task=COMMON_SENSE,
     languages=[DANISH],
     _labels=["a", "b"],
+    unofficial=True,
+)
+
+ZEBRA_PUZZLE_DA_CONFIG_2X3 = DatasetConfig(
+    name="zebra-puzzle-da-2x3",
+    pretty_name="the Danish version of the zebra puzzle dataset with 2x3 attributes "
+    "and 5 red herrings in the house theme.",
+    source="EuroEval/zebra-puzzles-da_huse_2x3_5rh",
+    task=LOGIC,
+    languages=[DANISH],
+    unofficial=True,
+)
+
+ZEBRA_PUZZLE_DA_CONFIG_4X5 = DatasetConfig(
+    name="zebra-puzzle-da-4x5",
+    pretty_name="the Danish version of the zebra puzzle dataset with 4x5 attributes "
+    "and 5 red herrings in the house theme.",
+    source="EuroEval/zebra-puzzles-da_huse_4x5_5rh",
+    task=LOGIC,
+    languages=[DANISH],
     unofficial=True,
 )